Removed HTML ID's, fixed cursor issue, UI issues
Removed all instances of HTML IDs being used in the javascript. This
will probably be added back once the installer is created though. This
allows the current form to be used multiple times on the same page.

Fixed an issue with the cursor not going to the correct position after
using the BBCode buttons. Changed the JS to use a simple implementation
that relies on the textNode created to figure out the cursor position.

Fixed a few lingering CSS issues in relation to scrolling text in both
editors once the size of the text got too large.

closes #10, #8