Medium Editor added again
Added the Medium Editor back to the Epochtalk-editor. Re-wrote the
angular code to use the Medium Editor and then parse BBCode on input and
blur events.

Removed the Buttons in the top nav and will be added a popover that
tells the user which bbcodes are accepted.

Added a bash script that will copy the Medium-Editor CSS over to app/css
since CSSify can't handle css from node_modules.