Exports directive to app param
Changed the way the default exports behavior works so that an existing
Angular app can be passed in and the directive will be attached to the
given app.

Moved all the code surrounding the directive in the main app.js file
into a testapp file. This include the Angular requires, the test parent
controller, and the test Angular app. Now the app.js file only requires
the CSS files for the editor and exports just a function that attaches
the directive.

Updated the README to reflect the changes above and added example code.

Updated the .gitignore for the moved files.